+++ setup.php
@@ -17,7 +17,7 @@
 use Glpi\Plugin\Hooks;
 use GlpiPlugin\Torah\Infrastructure\Glpi\HookBridge;
 
-define('PLUGIN_TORAH_VERSION', '0.3.0');
+define('PLUGIN_TORAH_VERSION', '0.3.1');
 define('PLUGIN_TORAH_MIN_GLPI_VERSION', '10.0.10');
 define('PLUGIN_TORAH_MAX_GLPI_VERSION', '10.0.99');
 define('PLUGIN_TORAH_MIN_PHP_VERSION', '8.2.0');
@@ -30,6 +30,7 @@ function plugin_init_torah(): void {
 
     $PLUGIN_HOOKS[Hooks::CSRF_COMPLIANT]['torah'] = true;
     $PLUGIN_HOOKS['config_page']['torah'] = 'front/config.php';
+    $PLUGIN_HOOKS[Hooks::JAVASCRIPT]['torah_admin_policy_matrix'] = ['plugins/torah/js/admin-policy-matrix.js'];
     $PLUGIN_HOOKS[Hooks::POST_ITEM_FORM]['torah'] = [HookBridge::class, 'postItemForm'];
     $PLUGIN_HOOKS[Hooks::FILTER_ACTORS]['torah'] = [HookBridge::class, 'filterActors'];